useSignaling hook
