+++ apps/example/src/components/Chat.tsx
@@ -3,12 +3,11 @@ import { useRtc, type EventHandler } from '@torolocos/react-rtc';
 import './styles.css';
 
 const Chat = () => {
-  const { send, enter, disconnect, state, on, off } = useRtc();
+  const { send, enter, disconnect, on, off } = useRtc();
   const [inputValue, setInputValue] = useState('');
   const [messageData, setMessageData] = useState([]);
   const [error, setError] = useState('');
-  const [, setChatOpen] = useState(false);
-  const { isEntered } = state;
+  const [isChatOpen, setChatOpen] = useState(false);
 
   const onStartChat = () => {
     enter(`${Math.random().toFixed(2)}`); //FIXME: tmp till we create valid Chat Screen
@@ -72,7 +71,7 @@ const Chat = () => {
           </div>
         ))}
       </div>
-      {isEntered && (
+      {isChatOpen && (
         <>
           <input
             value={inputValue}
@@ -81,8 +80,8 @@ const Chat = () => {
           <button onClick={onMessageSend}>send</button>
         </>
       )}
-      <button onClick={!isEntered ? onStartChat : onEndChat}>
-        {!isEntered ? 'join' : 'leave chat'}
+      <button onClick={!isChatOpen ? onStartChat : onEndChat}>
+        {!isChatOpen ? 'join' : 'leave chat'}
       </button>
     </div>
   );
